+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.2.0
+- Added a local chat message for when a player votes (configurable).
+- Votes will now be removed if their associated player disconnects.
+
 ## 1.1.0
 - Added a way open the Poll Creator menu by sending a phrase in chat (defaulted to `/poll`, configurable).
 - Added config options for automatically closing the Poll Creator menu after pressing the start/end poll button.